pyproject.toml
[project]
name = "image-watermarker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "briefcase>=0.3.24",
    "pillow>=11.3.0",
    "pyside6>=6.9.1",
]

# This project was generated with 0.3.24 using template: https://github.com/beeware/briefcase-template @ v0.3.24
[tool.briefcase]
project_name = "Python App"
bundle = "pro.srkdesign"
version = "0.0.1"
url = "https://srkdesign.pro/apps/image-watermarker"
license.file = "LICENSE"
author = "srkdesign"
author_email = "99723068+srkdesign@users.noreply.github.com"

[tool.briefcase.app.image_watermarker]
formal_name = "Python App"
description = "My first application"
long_description = """More details about the app should go here.
"""
sources = [
    "src/image_watermarker",
]
test_sources = [
    "tests",
]

requires = [
    "PySide6-Essentials~=6.8",
    # "PySide6-Addons~=6.8",
]
test_requires = [
    "pytest",
]

[tool.briefcase.app.image_watermarker.macOS]
universal_build = true
# As of Pyside 6.8, PySide enforces a macOS 12 minimum on wheels.
min_os_version = "12.0"
requires = [
    "std-nslog~=1.0.3",
]

[tool.briefcase.app.image_watermarker.linux]
requires = [
]

[tool.briefcase.app.image_watermarker.linux.system.debian]
system_requires = [
]

system_runtime_requires = [
    # Derived from https://doc.qt.io/qt-6/linux-requirements.html
    "libxext6",
    "libxrender1",
    "libx11-xcb1",
    "libxkbcommon-x11-0",
    "libxcb-image0",
    "libxcb-cursor0",
    "libxcb-shape0",
    "libxcb-randr0",
    "libxcb-xfixes0",
    "libxcb-sync1",
    "libxcb-icccm4",
    "libxcb-keysyms1",
    "libfontconfig1",
    "libsm6",
    "libice6",
    "libglib2.0-0",
    "libgl1",
    "libegl1",
    "libdbus-1-3",
]

[tool.briefcase.app.image_watermarker.linux.system.rhel]
system_requires = [
]

system_runtime_requires = [
    "qt6-qtbase-gui",
]

[tool.briefcase.app.image_watermarker.linux.system.suse]
system_requires = [
]

system_runtime_requires = [
    "libgthread-2_0-0",
    "libQt6Gui6",
]

[tool.briefcase.app.image_watermarker.linux.system.arch]
system_requires = [
]

system_runtime_requires = [
    "qt6-base",
]

[tool.briefcase.app.image_watermarker.linux.appimage]
manylinux = "manylinux_2_28"

system_requires = [
# ?? FIXME
]

linuxdeploy_plugins = [
]

[tool.briefcase.app.image_watermarker.linux.flatpak]
flatpak_runtime = "org.kde.Platform"
flatpak_runtime_version = "6.9"
flatpak_sdk = "org.kde.Sdk"

[tool.briefcase.app.image_watermarker.windows]
requires = [
]

# Mobile deployments
[tool.briefcase.app.image_watermarker.iOS]
supported = false

[tool.briefcase.app.image_watermarker.android]
supported = false

# Web deployments
[tool.briefcase.app.image_watermarker.web]
supported = false

